requirements.txt
numpy
pandas
requests
opencv-python
numpy
Pillow
transformers
torch